+++ src/form/types/types.ts
@@ -487,7 +487,12 @@ export interface IHeading3Field extends IFormFieldBase {
 export interface ISignatureFormField extends IFormFieldBase {
   type: typeof SIGNATURE
   maxSizeMb?: number
-  allowedFileFormats?: ('png' | 'jpg' | 'jpeg' | 'svg')[]
+  allowedFileFormats?: (
+    | 'image/png'
+    | 'image/jpg'
+    | 'image/jpeg'
+    | 'image/svg'
+  )[]
 }
 
 export type IFormField =